% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_cols_to_remove.R
\name{identify_cols_to_remove}
\alias{identify_cols_to_remove}
\title{Identifier les colonnes à supprimer / Identify Columns to Remove}
\usage{
identify_cols_to_remove(data)
}
\arguments{
\item{data}{\strong{Français :} Un data frame dans lequel les colonnes contenant uniquement des valeurs manquantes spécifiques
seront identifiées.
\strong{English :} A data frame in which columns containing only specified missing values will be identified.}
}
\value{
\strong{Français :} Un vecteur de noms de colonnes à supprimer, c'est-à-dire celles contenant uniquement
les valeurs manquantes spécifiées.
\strong{English :} A vector of column names to remove, i.e., those containing only the specified missing values.
}
\description{
\strong{Français :} Cette fonction identifie les colonnes dans un data frame contenant uniquement des valeurs
spécifiques de données manquantes (\code{-999999999}, \verb{##N/A##}, ou \code{NA}). Elle est conçue pour une utilisation
interne au package et n'est pas destinée à être directement accessible par les utilisateurs.

\strong{English :} This function identifies columns in a data frame that contain only specified missing values
(\code{-999999999}, \verb{##N/A##}, or \code{NA}). It is designed for internal package use and is not intended to be
directly accessible by users.
}
\details{
\strong{Français :} Cette fonction utilise \code{dplyr} pour sélectionner les colonnes numériques et de type caractère,
puis vérifie si elles ne contiennent que les valeurs \code{-999999999}, \verb{##N/A##}, ou \code{NA}. Les noms de colonnes
identifiés sont renvoyés sous forme de vecteur.

\strong{English :} This function uses \code{dplyr} to select numeric and character columns, then checks if they contain
only \code{-999999999}, \verb{##N/A##}, or \code{NA} values. The identified column names are returned as a vector.
}
\examples{
\dontrun{
# Exemple en français
cols_to_remove <- identify_cols_to_remove(data)

# English example
cols_to_remove <- identify_cols_to_remove(data)
}
}
