% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_metadata.R
\name{import_metadata}
\alias{import_metadata}
\title{Importer et structurer les métadonnées de Survey Solutions / Import and Tidy Survey Solutions Metadata}
\usage{
import_metadata(dossier)
}
\arguments{
\item{dossier}{\strong{Français :} Un caractère spécifiant le chemin vers le dossier contenant les fichiers \code{.do}
exportés depuis Survey Solutions, ainsi que des fichiers supplémentaires optionnels dans
\code{Questionnaire/content.zip}.
\strong{English :} A character string specifying the path to the folder containing \code{.do} files
exported from Survey Solutions and optional additional files in the \code{Questionnaire/content.zip}.}
}
\value{
\strong{Français :} Un data frame structuré (tidy) contenant les métadonnées, incluant les noms de variables,
leurs labels, ainsi que les labels associés aux valeurs.
\strong{English :} A tidy data frame with metadata, including variable names, labels, and associated value labels.
}
\description{
\strong{Français :} Cette fonction importe les métadonnées provenant des données exportées par Survey Solutions,
organise les labels de variables et de valeurs dans un format tidy, et remplace éventuellement
les labels existants par ceux provenant de fichiers de catégories supplémentaires, si disponibles.

\strong{English :} This function imports metadata from Survey Solutions exported data, organizes
variable and value labels in a tidy format, and optionally replaces existing labels with those from
additional category files if available.
}
\details{
\strong{Français :} La fonction traite les fichiers \code{.do} dans le dossier spécifié pour extraire les labels de
variables et de valeurs. Elle recherche également un fichier \code{content.zip} dans le sous-dossier
\code{Questionnaire} pour remplacer les labels par ceux des fichiers de catégories s'ils correspondent
aux labels de variables exportés.
\strong{English :} The function processes \code{.do} files in the specified folder to extract variable and value labels.
It also looks for a \code{content.zip} file within the \code{Questionnaire} subfolder to replace labels with those
from category files if they match the exported variable labels.
}
\examples{
\dontrun{
# Exemples d'utilisation / Example usage:
metadata <- import_metadata("chemin/vers/dossier/donnees")
head(metadata)
}

}
