% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_na_values.R
\name{replace_na_values}
\alias{replace_na_values}
\title{Remplacer les valeurs manquantes spécifiques par des NA / Replace Specific Missing Values with NA}
\usage{
replace_na_values(data)
}
\arguments{
\item{data}{\strong{Français :} Le data frame contenant les valeurs à remplacer.
\strong{English :} The data frame containing the values to replace.}
}
\value{
\strong{Français :} Un data frame avec les valeurs spécifiques de données manquantes remplacées par \code{NA}.
\strong{English :} A data frame with specific missing values replaced by \code{NA}.
}
\description{
\strong{Français :} Cette fonction détecte et remplace les valeurs spécifiques de données manquantes
(\verb{##N/A##} et \code{-999999999}) par des \code{NA} dans toutes les colonnes d'un jeu de données. Elle utilise
\code{dplyr} pour une application rapide et efficace, facilitant ainsi le traitement et l'analyse des données.

\strong{English :} This function detects and replaces specific missing values (\verb{##N/A##} and \code{-999999999})
with \code{NA} across all columns in a dataset. It leverages \code{dplyr} for fast and efficient application,
making data cleaning and analysis easier.
}
\details{
\strong{Français :} Cette fonction utilise \code{mutate(across(everything(), ...))} pour appliquer les changements
à toutes les colonnes. Elle remplace chaque occurrence de \verb{##N/A##} et \code{-999999999} par \code{NA}, ce qui permet
de standardiser les valeurs manquantes dans le jeu de données.

\strong{English :} This function uses \code{mutate(across(everything(), ...))} to apply changes across all columns.
It replaces each occurrence of \verb{##N/A##} and \code{-999999999} with \code{NA}, allowing for standardized missing values
throughout the dataset.
}
\examples{
\dontrun{
# Exemple en français
ESS_2024_results_Naclean <- ESS_2024_results_cleanlists \%>\%
  replace_na_values()

# English example
ESS_2024_results_Naclean <- ESS_2024_results_cleanlists \%>\%
  replace_na_values()
}

}
